+++ installation-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### DEV
 
+ * Check all innodb_large_prefix requirements in the install tool (see #80).
  * Use the table options instead of the default table options to compare engine and collation.
 
 ### 4.5.0-RC1 (2017-12-12)